version 3.8.01028

+++ app/src/main/java/com/getpoint/farminfomanager/entity/coordinate/LatLongAlt.java
@@ -39,7 +39,7 @@ public class LatLongAlt extends LatLong {
         return mAltitude;
     }
 
-    public void setAltitude(int altitude) {
+    public void setAltitude(float altitude) {
         this.mAltitude = altitude;
     }
